fix(state): move PendingIssueCompletion to state layer (#514)
The Architecture Layers CI check enforces "state must not depend on
TUI." The previous commit on this branch put the canonical
`PendingIssueCompletion` definition in `tui/app/types.rs` and let
`MaestroState::pending_completions` reach across the layer boundary.

Move the canonical definition to `state/types.rs` (next to
`MaestroState`); re-export from `tui/app/types.rs` so existing call
sites that use `super::types::PendingIssueCompletion` keep compiling.
The TUI layer depending on the state layer is allowed by the rule
(higher layer can use lower); the inverse is not.

`scripts/check-layers.sh` now passes locally.
